Moved MessageException to dataaccess
